requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
pydantic==2.10.4
python-dotenv==1.0.1
google-api-python-client==2.159.0
google-auth==2.37.0
pytest==8.3.4
httpx==0.28.1

